Completion command for scopes inside language grammars works mostly identical to the standard completion with ⎋. However you will not get a beep when there are no options, it will just silently exit. Hopefully allan gives us a new exit code for producing a beep since the alternatives suck. ;)
Note that a lot of the completions in this are just guidelines, use the built in validation to see where custom words are allowed (or the documentation with ⌃h).

In the future this will hopefully also work for scope fields in various windows but that will have to wait for allan to give up better extensions that just .textmate when inside TextMate using Edit in TextMate. :)


git-svn-id: http://svn.textmate.org/trunk/Bundles/TextMate.tmbundle@6187 dfb7d73b-c2ec-0310-8fea-fb051d288c6d