• allow pasting to be aborted
Conceptually this was just giving a “kill curl” lambda block to the call_with_progress, but effectively turned out to be more involved, since any shell execution done in the call_with_progress block caused the cancel lambda block to not to be called, so no calling out to `mktemp`, and the find_ext was indirectly calling `ps` (to find the TM app path). So I had to change these things.

Additionally `curl` is no longer called with %x{…} (since we need a pid to kill, well, actually, I made it kill all processes with pgid matching the current pid, since at the time of providing the cancel block, curl’s pid is not known) — but since the result from curl is needed, handling curl’s stdout (by creating a pipe) was required…

Anyway, it works :) Now we just need some graceful error reporting for when pasting fails.


git-svn-id: http://svn.textmate.org/trunk/Bundles/TextMate.tmbundle@8569 dfb7d73b-c2ec-0310-8fea-fb051d288c6d